+++ test/ModuleTest.php
@@ -2,6 +2,7 @@
 namespace AcMailerTest;
 
 use AcMailer\Module;
+use AcMailerTest\Console\AdapterMock;
 use PHPUnit_Framework_TestCase as TestCase;
 
 /**
@@ -27,5 +28,16 @@ class ModuleTest extends TestCase
         $returnedConfig = $this->module->getConfig();
 
         $this->assertEquals($expectedConfig, $returnedConfig);
+    }
+
+    public function testGetConsoleUsage()
+    {
+        $this->assertEquals(
+            [
+                'acmailer parse-config [--configKey=] [--format=(php|xml|ini|json)] [--outputFile=]'
+                => 'Parses the configuration of AcMailer module <=v4.5.1 to the structure used in v5.0.0'
+            ],
+            $this->module->getConsoleUsage(new AdapterMock())
+        );
     }
 }